requirements.txt
actionlib==1.14.0
angles==1.9.13
attrs==22.2.0
bokeh==3.0.3
bondpy==1.8.6
camera-calibration==1.17.0
camera-calibration-parsers==1.12.0
catkin==0.8.10
catkin-pkg==0.5.2
certifi==2022.12.7
chardet==5.1.0
charset-normalizer==3.1.0
click==8.1.3
colorlog==6.7.0
contourpy==1.0.7
controller-manager==0.19.6
controller-manager-msgs==0.19.6
cv-bridge==1.16.2
cycler==0.11.0
Cython==0.29.33
descartes==1.1.0
diagnostic-analysis==1.11.0
diagnostic-common-diagnostics==1.11.0
diagnostic-updater==1.11.0
distro==1.8.0
docutils==0.19
dynamic-reconfigure==1.7.3
fonttools==4.39.0
gazebo-plugins==2.9.2
gazebo-ros==2.9.2
gencpp==0.7.0
geneus==3.0.0
genlisp==0.4.18
genmsg==0.6.0
gennodejs==2.0.2
genpy==0.6.15
idna==3.4
image-geometry==1.16.2
imageio==2.26.0
importlib-resources==5.12.0
interactive-markers==1.12.0
Jinja2==2.10.3
joint-state-publisher==1.15.1
joint-state-publisher-gui==1.15.1
jsonschema==4.17.3
kiwisolver==1.4.4
laser-geometry==1.6.7
lazy-loader==0.1
lxml==4.9.2
mapbox-earcut==1.0.1
MarkupSafe==2.0.1
matplotlib==3.7.1
message-filters==1.15.15
mpmath==1.2.1
msgpack==1.0.4
networkx==3.0
noise==1.2.2
numpy==1.24.2
packaging==23.0
pandas==1.5.3
pcg-gazebo==0.7.12
Pillow==9.4.0
pkgutil-resolve-name==1.3.10
progress==1.6
psutil==5.9.4
pycollada==0.6
pyglet==1.5.27
pyparsing==3.0.9
pyrsistent==0.19.3
python-dateutil==2.8.2
python-fcl==0.7.0.4
python-qt-binding==0.4.4
pytoml==0.1.21
pytz==2022.7.1
PyWavelets==1.4.1
PyYAML==6.0
qt-dotgraph==0.4.2
qt-gui==0.4.2
qt-gui-cpp==0.4.2
qt-gui-py-common==0.4.2
requests==2.28.2
resource-retriever==1.12.7
rosbag==1.15.15
rosboost-cfg==1.15.8
rosclean==1.15.8
roscreate==1.15.8
rosgraph==1.15.15
roslaunch==1.15.15
roslib==1.15.8
roslint==0.12.0
roslz4==1.15.15
rosmake==1.15.8
rosmaster==1.15.15
rosmsg==1.15.15
rosnode==1.15.15
rosparam==1.15.15
rospkg==1.4.0
rospy==1.15.15
rosservice==1.15.15
rostest==1.15.15
rostopic==1.15.15
rosunit==1.15.8
roswtf==1.15.15
rqt-action==0.4.9
rqt-bag==0.5.1
rqt-bag-plugins==0.5.1
rqt-console==0.4.11
rqt-dep==0.4.12
rqt-graph==0.4.14
rqt-gui==0.5.3
rqt-gui-py==0.5.3
rqt-image-view==0.4.16
rqt-launch==0.4.9
rqt-logger-level==0.4.11
rqt-moveit==0.5.10
rqt-msg==0.4.10
rqt-nav-view==0.5.7
rqt-plot==0.4.13
rqt-pose-view==0.5.11
rqt-publisher==0.4.10
rqt-py-common==0.5.3
rqt-py-console==0.4.10
rqt-reconfigure==0.5.5
rqt-robot-dashboard==0.5.8
rqt-robot-monitor==0.5.14
rqt-robot-steering==0.5.12
rqt-runtime-monitor==0.5.9
rqt-rviz==0.7.0
rqt-service-caller==0.4.10
rqt-shell==0.4.11
rqt-srv==0.4.9
rqt-tf-tree==0.6.3
rqt-top==0.4.10
rqt-topic==0.4.13
rqt-web==0.4.10
Rtree==1.0.1
rviz==1.14.19
scikit-image==0.20.0
scipy==1.9.1
sensor-msgs==1.13.1
Shapely==1.7.0
six==1.16.0
smach==2.5.0
smach-ros==2.5.0
smclib==1.8.6
svg.path==6.2
sympy==1.11.1
tabulate==0.9.0
tf==1.13.2
tf-conversions==1.13.2
tf2-geometry-msgs==0.7.6
tf2-kdl==0.7.6
tf2-py==0.7.6
tf2-ros==0.7.6
tifffile==2023.2.28
topic-tools==1.15.15
tornado==6.2
triangle==20220202
trimesh==3.20.1
urllib3==1.26.14
xacro==1.14.14
xmltodict==0.13.0
xxhash==3.2.0
xyzservices==2023.2.0
yasha==5.0
zipp==3.15.0
